Bugfixes for core dumps during backup jobs.
Remove the call to io_workers_stop() in pfs_rpc_close() (called from
proxyfs_unmount()) because there can be multiple mounts by a single Samba
process if the SMB client requests it by opening multiple sessions in
the same TCP connection, which it can do.

Add a mutex implementation in synchron.[hc] that does error checking
and panics if there's a problem.  It is not yet called anywhere.

Clean up some other obvious issues or code in the jrpcclient sources.