Rework the code that handles disconnects.
The current code in jsonrpc_rpc_bounce() has a number of flaws, among
them:
  o after the old socket pool is destroyed and a new socket pool
    created, existing open sockets are returned to the new pool, which
    they are not part of; this might cause the same socket to be given
    out twice
  o if a thread is waiting for a socket when the connection pool
    is destroyed it may never wake up
  o sockets in the destroyed connection pool are never closed, which
    may cause us to leak file descriptors (but probably avoids the
    same fd being given out twice).

Change the socket pool code to add a sock_pool_put_badfd() function that
puts the socket back into the pool, waits for all sockets to be idle,
closes them, and then re-opens them.

Get rid of jsonrpc_rpc_bounce() and the call to it in
jsonrpc_free_read_buf().  Any socket errors now trigger all sockets to
be re-opened when the error is detected.

Revise sock_write() so that all error reporting goes through a single
section of code (at the errout label).  If we succesfully got a socket
but it turns out to be "broken", we still need to return it to the
global pool.