Improved Readdir*() JSON RPCs to begin to allow for MaxEntries > 1
The current version of jrpcclient mandates a return of zero or one
response entries... and also that a zero response results in a
return status of ENOENT. This is no longer the behavior or ProxyFS,
so ENOENT is locally generated when the length of the response
(i.e. # of returned dirEntries) is zero.